// src/core/enums/commodity.enum.ts
export enum Commodity {
  STUFF_INIT = 0,
  STUFF_FOOD = 0,
  STUFF_JOBS = 1,
  STUFF_COAL = 2,
  STUFF_GOODS = 3,
  STUFF_ORE = 4,
  STUFF_STEEL = 5,
  STUFF_WASTE = 6,
  STUFF_KWH = 7,
  STUFF_MWH = 8,
  STUFF_WATER = 9,
  STUFF_CRICKET = 10,
  STUFF_FIRE = 11,
  STUFF_HEALTH = 12,
  STUFF_MAX_FOOD = 13,
  STUFF_MAX_JOBS = 14,
  STUFF_MAX_COAL = 15,
  STUFF_MAX_GOODS = 16,
  STUFF_MAX_ORE = 17,
  STUFF_MAX_STEEL = 18,
  STUFF_MAX_WASTE = 19,
  STUFF_COUNT = 20
}